+++ src/lexer/lexer.ts
@@ -176,6 +176,7 @@ export class Lexer {
 		yield* this.match(context, [
 			[MATCH_XML_TAG, State.INITIAL, false],
 			[MATCH_DOCTYPE_OPEN, State.DOCTYPE, TokenType.DOCTYPE_OPEN],
+			[MATCH_WHITESPACE, State.INITIAL, TokenType.WHITESPACE],
 			[false, State.TEXT, false],
 		], 'expected doctype');
 	}